Fixed a bug.
Former-commit-id: 95bf9a32fbcfea2fe47ff93fe87467f95d9e01c8

+++ PC/IPIC3D2/src/particles/Particles3Dcomm.cpp
@@ -1423,7 +1423,7 @@ void Particles3Dcomm::recommunicate_particles_until_done(int min_num_iterations)
   // * to minimize unnecessary sends,
   // with the overall goal of minimizing time spent in communication
   //
-  int num_pcls_sent;
+  long long num_pcls_sent;
   for(int i=0;i<min_num_iterations;i++)
   {
     flush_send(); // flush sending of particles
